chore: Implement navigation structure with bottom tabs and routes
- Added AppNavigationRoutes.kt to define navigation routes and bottom tab structure.
- Introduced BottomTab enum for managing tab routes, icons, and labels.
- Created AppNavigationScaffold.kt to manage the main navigation scaffold with bottom navigation.
- Implemented NavigationTabContainer for handling tab transitions and visibility.
- Added AppNavigationRoutesTest.kt to ensure correct route detection and bottom bar visibility.
